requirements.txt
streamlit
chromadb
PyPDF2
sentence-transformers
langchain
langchain-community
litellm
huggingface-hub
python-dotenv
